refactor: unit-util 클래스화
