+++ src/unit-util/unit-converters/byte-unit-converter.type.ts
@@ -1,0 +1,1 @@
+export type ByteUnitType = 'bytes' | 'kilobytes' | 'megabytes';
