+++ src/unit-util/unit-converters/index.ts
@@ -1,0 +1,4 @@
+import { ByteUnitConverter } from './byte-unit-converter';
+
+export type { ByteUnitType } from './byte-unit-converter.type';
+export const byteUnitConverter = new ByteUnitConverter();
