--- src/unit-util/unit-util.type.ts
@@ -1,2 +1,0 @@
-export type DataSizeType = 'bytes' | 'kilobytes' | 'megabytes';
-export type DataSizeNotationMapType = Record<DataSizeType, string>;
